# Optimize Enumerable.Range(...).Select(...)