@stephentoub 2019-05-12
I found very little usage of Repeat in general outside of test code, and very little of that used Select. What's the scenario where it's important that be a fast as possible?